fix(runtime): prevent uint32_t multiplication overflow in adapter router and HFT gate
adapter.c: Widen `used * 100u` to uint64_t before dividing by
scaled_capacity, preventing overflow when `used` exceeds ~42 million
(0xFFFFFFFF / 100).

hft_instrument.c: Same fix for p99.9 and p99.99 percentile target
calculations — `hist->total * 999u` and `hist->total * 9999u` both
overflow uint32_t when total exceeds ~4.3M and ~430K respectively.
Cast to uint64_t before the multiply.

+++ src/runtime/adapter.c
@@ -220,7 +220,7 @@ void asx_adapter_router_decide(uint32_t used, uint32_t capacity,
         return;
     }
 
-    load_pct = (used * 100u) / scaled_capacity;
+    load_pct = (uint32_t)((uint64_t)used * 100u / scaled_capacity);
     out->load_pct = load_pct;
 
     if (load_pct >= 75) {
